feat(active-active): compressing-cluster-attr-with-snappy (#7371)
<!-- Describe what has changed in this PR -->
**What changed?**
switches the Cluster attributes as they're stored in the DB to use
snappy, since they're expected to potentially be somewhat large.

I expect this to be a backwards compatible change due to the existing
records being persisted with the encoding type, and that's what local
testing showed to be true.

<!-- Tell your future self why have you made these changes -->
**Why?**

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
- [X] Tested manually
- [X] Unit tests

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Signed-off-by: David Porter <david.porter@uber.com>
Signed-off-by: Gaziza Yestemirova <gaziza@uber.com>