feat: Updating warning for UpdateDomain cli command (#7375)
<!-- Describe what has changed in this PR -->
**What changed?**

The UpdateDomain API is (optionally) undergoing changes where it's
administrative functions are being protected by authorization checks and
it's failover / user-facing functionality is being made available in the
new endpoint `Failover`. This - if enforced - may affect users who're
not aware of this change.

<!-- Tell your future self why have you made these changes -->
**Why?**

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Signed-off-by: David Porter <david.porter@uber.com>
Signed-off-by: Gaziza Yestemirova <gaziza@uber.com>