feat: Add domain audit table (#7376)
<!-- Describe what has changed in this PR -->
**What changed?**

Adds the domain_audit_log table.

<!-- Tell your future self why have you made these changes -->
**Why?**

This is the first step of the persistence implementation for a persisted
audit_log for domain changes. It will only be used for changes to the
ReplicationConfig initially, replacing FailoverHistory in the domains
metadata.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**

Unit tests, manual POC.

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

Something is wrong with the schema definition (or the primary key
definition) and we need to modify the table later. Ideally this is
caught before it gets much further than this.

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

N/A

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

N/A

**Detailed Description**
[In-depth description of the changes made to the schema or interfaces,
specifying new fields, removed fields, or modified data structures]

The domain_audit_log table has been added to the Cassandra schema. It
will not exist in SQL etc. for now, and is planned to be added early
next year.

**Impact Analysis**
- **Backward Compatibility**: [Analysis of backward compatibility]
- **Forward Compatibility**: [Analysis of forward compatibility]

N/A

**Testing Plan**
- **Unit Tests**: [Do we have unit test covering the change?]
- **Persistence Tests**: [If the change is related to a data type which
is persisted, do we have persistence tests covering the change?]
- **Integration Tests**: [Do we have integration test covering the
change?]
- **Compatibility Tests**: [Have we done tests to test the backward and
forward compatibility?]

This should be covered by future integration & persistence tests - but
is not yet. They will be added in a follow up PR.

**Rollout Plan**
- What is the rollout plan?
- Does the order of deployment matter? No
- Is it safe to rollback? Does the order of rollback matter? Yes, until
applications start using it.
- Is there a kill switch to mitigate the impact immediately? No.
---

Signed-off-by: Gaziza Yestemirova <gaziza@uber.com>