fix(shard-distributor): remove trimming of prefixes (#7490)
<!-- Describe what has changed in this PR -->
**What changed?**
Reverting the trimprefix since we are using constants to compare the
values that include that

<!-- Tell your future self why have you made these changes -->
**Why?**
Constants that include the prefix are used to

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Deployed in staging

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
Corruption of db, which is already the case.

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: edigregorio <edigregorio@uber.com>
Signed-off-by: Gaziza Yestemirova <gaziza@uber.com>